requirements.txt
fastapi
uvicorn[standard]
python-jose
bcrypt
pydantic-settings
supabase
